Moved merge_land_and_sea function to calibration utilities
